sms-gammu-gateway/requirements.txt
python-gammu
flask
flask-restx
Flask-HTTPAuth
pyopenssl
paho-mqtt
click